services/recommender/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bson==0.5.10
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
joblib==1.3.2
mongoengine==0.27.0
motor==3.3.1
numpy==1.26.2
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
uvicorn==0.23.2
